Move example code to _example
