+++ _example/go.mod
@@ -1,0 +1,27 @@
+module example
+
+go 1.18
+
+require (
+	github.com/oklahomer/go-sarah-githubconfig v0.0.0-20220611010143-046d230dcbb9
+	github.com/oklahomer/go-sarah/v4 v4.0.3
+)
+
+require (
+	github.com/golang/protobuf v1.5.2 // indirect
+	github.com/gorilla/websocket v1.4.2 // indirect
+	github.com/oklahomer/go-kasumi v0.0.0-20220203122045-3db87696aa9c // indirect
+	github.com/oklahomer/golack/v2 v2.1.0 // indirect
+	github.com/patrickmn/go-cache v2.1.0+incompatible // indirect
+	github.com/robfig/cron/v3 v3.0.1 // indirect
+	github.com/shurcooL/githubv4 v0.0.0-20220520033151-0b4e3294ff00 // indirect
+	github.com/shurcooL/graphql v0.0.0-20220606043923-3cf50f8a0a29 // indirect
+	github.com/tidwall/gjson v1.10.1 // indirect
+	github.com/tidwall/match v1.1.1 // indirect
+	github.com/tidwall/pretty v1.2.0 // indirect
+	golang.org/x/net v0.0.0-20220607020251-c690dde0001d // indirect
+	golang.org/x/oauth2 v0.0.0-20220608161450-d0670ef3b1eb // indirect
+	google.golang.org/appengine v1.6.7 // indirect
+	google.golang.org/protobuf v1.28.0 // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
+)
