+++ _example/plugin/hello/command.go
